fix: move StyledBottom component to common component
